pyproject.toml
[project]
name = "retrai"
version = "0.1.0"
description = "Self-solving AI agent loop — run a goal, watch it fix itself"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.5.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "litellm>=1.50.0",
    "typer>=0.15.0",
    "textual>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "pytest-json-report>=1.5.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pyright>=1.1.389",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
]

[project.scripts]
retrai = "retrai.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["retrai"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
